started putting all prototypes into a real structure.
