+++ test/gui/lineplot.py
@@ -9,133 +9,6 @@ from plottr.data.datadict import DataDict
 from plottr.gui.ui_lineplot import Ui_LinePlot
 
 
-class NodeBase:
-
-    dataProcessed = pyqtSignal(object)
-    optionsUpdated = pyqtSignal()
-    dataRequested = pyqtSignal()
-
-    sources = ['input']
-
-    def __init__(self, *arg, **kw):
-        self.verbose = kw.get('verbose', False)
-
-        self._sources = {}
-        self._updateOnSource = True
-        self._uptodate = False
-
-        self._sources = { s : {'ref' : None, 'data' : None}  for s in self.__class__.sources }
-
-        self.optionsUpdated.connect(self.run)
-
-        for k, v in self._sources.items():
-            setattr(self, 'set' + k.capitalize(),
-                lambda ref: self.setSource(k, ref))
-            setattr(self, 'set' + k.capitalize() + 'Data',
-                pyqtSlot(object)(lambda data: self.setSourceData(k, data)))
-            setattr(self, 'get' + k.capitalize() + 'Data',
-                lambda: self._sources[k]['data'])
-
-    @staticmethod
-    def updateOption(func):
-        def wrap(self, val):
-            ret = func(self, val)
-            self._uptodate = False
-            self.optionsUpdated.emit()
-            return ret
-        return wrap
-
-    def setSourceData(self, sourceName, value):
-        if sourceName not in self._sources:
-            raise ValueError("'{}' is not a recognized source.")
-
-        if self.verbose:
-            print(self, 'data set from', sourceName)
-
-        self._sources[sourceName]['data'] = value
-        if self._updateOnSource:
-            self.run()
-        else:
-            self._uptodate = False
-
-    def setSource(self, sourceName, ref):
-        if sourceName not in self._sources:
-            raise ValueError("'{}' is not a recognized source." )
-
-        setfunc = getattr(self, 'set' + sourceName.capitalize() + 'Data')
-        src = self._sources[sourceName]['ref']
-        if src is not None:
-            src.dataProcessed.disconnect(setfunc)
-            self.dataRequested.disconnect(src.broadcastData)
-
-        self._sources[sourceName]['ref'] = ref
-        src = self._sources[sourceName]['ref']
-        src.dataProcessed.connect(setfunc)
-        self.dataRequested.connect(src.broadcastData)
-        self.dataRequested.emit()
-
-    def run(self):
-        raise NotImplementedError
-
-
-class Node(QObject, NodeBase):
-
-    def __init__(self):
-        super().__init__()
-        self._data = None
-        self._updateOnOptionChange = True
-        self._running = False
-
-    @property
-    def updateOnSource(self):
-        return self._updateOnSource
-
-    @updateOnSource.setter
-    def updateOnSource(self, val):
-        self._updateOnSource = val
-
-    @property
-    def updateOnOptionChange(self):
-        return self._updateOnOptionChange
-
-    @updateOnOptionChange.setter
-    def updateOnOptionChange(self, val):
-        self._updateOnOptionChange = val
-
-    @property
-    def data(self):
-        return self._data
-
-    @data.setter
-    def data(self, val):
-        self._data = val
-        self.broadcastData()
-
-    @pyqtSlot()
-    def broadcastData(self):
-        self.dataProcessed.emit(self._data)
-
-    @pyqtSlot()
-    def run(self):
-        if not self._running:
-            self._running = True
-            for n, s in self._sources.items():
-                if s['ref'] is not None:
-                    s['ref'].run()
-
-            if not self._uptodate:
-                self.data = self.processData()
-                if self.verbose:
-                    print('processed data:', self)
-                    pprint(self.data)
-                    print('')
-                self._uptodate = True
-                self._running = False
-
-    def processData(self):
-        return self.data
-
-
 class DataSelector(Node):
 
     def __init__(self):
